Updates session expiration and JS handlers
`handler.js` includes an IIFE at the end of the file, which is also an
example of one of the 4 situations where JS's ASI does not treat a \n as
a statement termination. This adds a semicolon before the IIFE to
indicate the special case.

See
[here](http://blog.izs.me/post/2353458699/an-open-letter-to-javascript-leaders-regarding) for more on this.

Session variable set to expire after a minute to acommodate with new
user locations during the course of a real-world session.